Fix: separa authOptions em arquivo separado e ajusta route.ts para App Router NextAuth
